Clean up HTML Area classes/code #720
-CKE postfix removed
-tinymce related files removed
-styles cleaned
