+++ src/main/resources/admin/tools/main/main.html
@@ -45,7 +45,6 @@
     var CONFIG = {
         appId: '{{appId}}',
         xpVersion: '{{xpVersion}}',
-        isCkeUsed: {{isCkeUsed}},
         appIconUrl: '{{assetsUri}}/icons/icon-white.svg',
         assetsUri: '{{assetsUri}}',
         adminUrl: '{{adminUrl}}',
