+++ src/main/resources/assets/js/app/wizard/PageComponentsView.ts
@@ -9,7 +9,6 @@ import {ItemViewSelectedEvent} from '../../page-editor/ItemViewSelectedEvent';
 import {ItemViewDeselectedEvent} from '../../page-editor/ItemViewDeselectedEvent';
 import {ComponentAddedEvent} from '../../page-editor/ComponentAddedEvent';
 import {TextComponentView} from '../../page-editor/text/TextComponentView';
-import {TextComponentViewCK} from '../../page-editor/text/TextComponentViewCK';
 import {FragmentComponentView} from '../../page-editor/fragment/FragmentComponentView';
 import {LayoutComponentView} from '../../page-editor/layout/LayoutComponentView';
 import {ComponentRemovedEvent} from '../../page-editor/ComponentRemovedEvent';
@@ -225,8 +224,7 @@ export class PageComponentsView
                             }
                         }
 
-                        if (api.ObjectHelper.iFrameSafeInstanceOf(event.getComponentView(), TextComponentView) ||
-                            api.ObjectHelper.iFrameSafeInstanceOf(event.getComponentView(), TextComponentViewCK)) {
+                        if (api.ObjectHelper.iFrameSafeInstanceOf(event.getComponentView(), TextComponentView)) {
                             this.bindTreeTextNodeUpdateOnTextComponentModify(<TextComponentView>event.getComponentView());
                         }
 
@@ -322,8 +320,7 @@ export class PageComponentsView
             }
 
             let clickedItemView: ItemView = this.tree.getGrid().getDataView().getItem(data.row).getData();
-            let isTextComponent = api.ObjectHelper.iFrameSafeInstanceOf(clickedItemView, TextComponentView) ||
-                                  api.ObjectHelper.iFrameSafeInstanceOf(clickedItemView, TextComponentViewCK);
+            let isTextComponent = api.ObjectHelper.iFrameSafeInstanceOf(clickedItemView, TextComponentView);
 
             if (isTextComponent) {
                 this.editTextComponent(clickedItemView);
@@ -435,8 +432,7 @@ export class PageComponentsView
         this.tree.getGrid().getDataView().getItems().map((dataItem) => {
             return dataItem.getData();
         }).filter((itemView: ItemView) => {
-            return api.ObjectHelper.iFrameSafeInstanceOf(itemView, TextComponentView) ||
-                   api.ObjectHelper.iFrameSafeInstanceOf(itemView, TextComponentViewCK);
+            return api.ObjectHelper.iFrameSafeInstanceOf(itemView, TextComponentView);
         }).filter((textComponentView: TextComponentView) => {
             return !textComponentView.getHTMLElement().onpaste; // filtering text components that already have these listeners
         }).forEach((textComponentView: TextComponentView) => {
