+++ src/main/resources/assets/page-editor/styles/partials/text-view.less
@@ -2,7 +2,7 @@
 
   overflow: auto;
 
-  // tiny mce style enforcing
+  // cke style enforcing
   ul, ol {
     margin: 0 0 12px 40px;
   }
@@ -21,7 +21,7 @@
     font-weight: inherit;
   }
 
-  // end of tiny mce style enforcing
+  // end of cke style enforcing
 
   &.empty .@{_CLS_PREFIX}text-placeholder {
     margin-bottom: 0;
@@ -32,33 +32,14 @@
     display: none;
   }
 
-  .mce-content-body {
-    cursor: auto;
-    display: none;
-    padding: 5px;
-
-    &.mce-edit-focus {
-      outline: none;
-    }
-  }
-
   .cke_editable {
     cursor: auto;
     display: none;
     padding: 5px;
-
-    &.mce-edit-focus {
-      outline: none;
-    }
   }
 
   &.edit-mode {
     min-height: 0;
-
-    .mce-content-body {
-      display: block;
-      box-sizing: border-box;
-    }
 
     .cke_editable {
       display: block;
@@ -116,69 +97,13 @@
   .icon-font-size;
 }
 
-.mce-toolbar-container {
-  position: fixed;
-  top: 0;
-  left: 0;
-  z-index: @z-index-mce-toolbar;
-  width: 100%;
-  border-bottom: 1px solid #cccccc;
-
-  > div {
-    display: none;
-  }
-  &.visible {
-    box-shadow: 0 8px 8px 0 @admin-toolbar-shadow;
-
-    > div {
-      display: block;
-    }
-  }
-
-  .mce-abs-layout, .mce-toolbar-grp, .mce-flow-layout, .mce-container-body.mce-stack-layout {
-    width: auto !important;
-  }
-
-  .mce-toolbar-grp, .mce-tinymce-inline, .mce-container-body.mce-stack-layout, .mce-container-body.mce-abs-layout {
-    height: auto !important;
-  }
-
-  .mce-container-body.mce-abs-layout {
-    overflow: initial;
-  }
-
-  .mce-abs-layout-item, .mce-abs-end {
-    position: static;
-  }
-
-  .mce-tinymce-inline {
-
-    width: 100% !important;
-
-    .mce-container-body {
-      margin: auto;
-    }
-
-    .mce-flow-layout {
-      white-space: normal;
-      padding-right: 30px;
-    }
-  }
-
-  .mce-toolbar-grp.mce-panel {
-    padding-bottom: 1px;
-    border: none;
-  }
-
-}
-
 .cke-toolbar-container {
   &.visible .cke_top {
     display: block !important;
   }
 
   .cke_combopanel {
-    z-index: @z-index-mce-toolbar !important;
+    z-index: @z-index-cke-toolbar !important;
   }
 }
 
@@ -186,7 +111,7 @@
   position: fixed !important;
   top: 0;
   left: 0;
-  z-index: @z-index-mce-toolbar;
+  z-index: @z-index-cke-toolbar;
   width: 100% !important;
   border-bottom: 1px solid #cccccc !important;
   text-align: center !important;
@@ -224,7 +149,7 @@
   top: 12px;
   right: 10px;
   line-height: 12px;
-  z-index: @z-index-mce-toolbar-close;
+  z-index: @z-index-cke-toolbar-close;
   pointer-events: auto;
   color: @admin-dark-gray;
   display: none;
@@ -239,8 +164,4 @@
   &:hover {
     cursor: pointer;
   }
-}
-
-.mce-grid-border a {
-  pointer-events: auto;
 }
