+++ src/main/resources/assets/page-editor/styles/variables.less
@@ -4,3 +4,7 @@
 
 @region-bg: #fff;
 @region-color: #add8e6;
+
+@z-index-cke-shader: 65535;
+@z-index-cke-toolbar-close: @z-index-cke-shader - 1;
+@z-index-cke-toolbar: @z-index-cke-toolbar-close - 1;
